Add profile for selective HDFS ACL support
