+++ underfs/hdfs/src/main/java/alluxio/underfs/hdfs/HdfsUnderFileSystem.java
@@ -17,8 +17,6 @@ import alluxio.PropertyKey;
 import alluxio.retry.CountingRetry;
 import alluxio.retry.RetryPolicy;
 import alluxio.security.authorization.AccessControlList;
-import alluxio.security.authorization.AclAction;
-import alluxio.security.authorization.AclEntryType;
 import alluxio.underfs.AtomicFileOutputStream;
 import alluxio.underfs.AtomicFileOutputStreamCallback;
 import alluxio.underfs.BaseUnderFileSystem;
@@ -46,12 +44,8 @@ import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.fs.permission.AclEntry;
-import org.apache.hadoop.fs.permission.AclStatus;
-import org.apache.hadoop.fs.permission.FsAction;
 import org.apache.hadoop.fs.permission.FsPermission;
 import org.apache.hadoop.hdfs.DistributedFileSystem;
-import org.apache.hadoop.hdfs.protocol.AclException;
 import org.apache.hadoop.security.SecurityUtil;
 import org.apache.hadoop.security.UserGroupInformation;
 import org.slf4j.Logger;
@@ -82,6 +76,7 @@ public class HdfsUnderFileSystem extends BaseUnderFileSystem
   private static final String HDFS_USER = "";
 
   private final LoadingCache<String, FileSystem> mUserFs;
+  private final HdfsAclProvider mHdfsAclProvider;
   private UnderFileSystemConfiguration mUfsConf;
 
   /**
@@ -107,6 +102,24 @@ public class HdfsUnderFileSystem extends BaseUnderFileSystem
   public HdfsUnderFileSystem(AlluxioURI ufsUri, UnderFileSystemConfiguration conf,
       Configuration hdfsConf) {
     super(ufsUri, conf);
+
+    // Create the supported HdfsAclProvider if possible.
+    HdfsAclProvider hdfsAclProvider = new NoopHdfsAclProvider();
+    try {
+      Object o = Class.forName("alluxio.underfs.hdfs.acl.SupportedHdfsAclProvider").newInstance();
+      if (o instanceof HdfsAclProvider) {
+        hdfsAclProvider = (HdfsAclProvider) o;
+      } else {
+        LOG.warn(
+            "SupportedHdfsAclProvider is not instance of HdfsAclProvider. HDFS ACLs will not be "
+                + "supported.");
+      }
+    } catch (Exception e) {
+      // ignore
+      LOG.warn("Cannot create SupportedHdfsAclProvider. HDFS ACLs will not be supported.");
+    }
+    mHdfsAclProvider = hdfsAclProvider;
+
     mUfsConf = conf;
     Path path = new Path(ufsUri.toString());
     // UserGroupInformation.setConfiguration(hdfsConf) will trigger service loading.
@@ -227,50 +240,7 @@ public class HdfsUnderFileSystem extends BaseUnderFileSystem
 
   @Override
   public AccessControlList getAcl(String path) throws IOException {
-    FileSystem hdfs = getFs();
-    // TODO(chen): for hadoop <= 2.3, ACL is not supported, there is no getAclStatus method.
-    AclStatus hdfsAcl;
-    try {
-      hdfsAcl = hdfs.getAclStatus(new Path(path));
-    } catch (AclException e) {
-      // When dfs.namenode.acls.enabled is false, getAclStatus throws AclException.
-      return null;
-    }
-    AccessControlList acl = new AccessControlList();
-    acl.setOwningUser(hdfsAcl.getOwner());
-    acl.setOwningGroup(hdfsAcl.getGroup());
-    for (AclEntry entry : hdfsAcl.getEntries()) {
-      // TODO(chen): handle the case where the entry is for default ACL in a directory.
-      alluxio.security.authorization.AclEntry.Builder builder =
-          new alluxio.security.authorization.AclEntry.Builder();
-      builder.setType(getAclEntryType(entry));
-      builder.setSubject(entry.getName());
-      FsAction permission = entry.getPermission();
-      if (permission.implies(FsAction.READ)) {
-        builder.addAction(AclAction.READ);
-      } else if (permission.implies(FsAction.WRITE)) {
-        builder.addAction(AclAction.WRITE);
-      } else if (permission.implies(FsAction.EXECUTE)) {
-        builder.addAction(AclAction.EXECUTE);
-      }
-      acl.setEntry(builder.build());
-    }
-    return acl;
-  }
-
-  private AclEntryType getAclEntryType(AclEntry entry) throws IOException {
-    switch (entry.getType()) {
-      case USER:
-        return entry.getName().isEmpty() ? AclEntryType.OWNING_USER : AclEntryType.NAMED_USER;
-      case GROUP:
-        return entry.getName().isEmpty() ? AclEntryType.OWNING_GROUP : AclEntryType.NAMED_GROUP;
-      case MASK:
-        return AclEntryType.MASK;
-      case OTHER:
-        return AclEntryType.OTHER;
-      default:
-        throw new IOException("Unknown ACL entry type: " + entry.getType());
-    }
+    return mHdfsAclProvider.getAcl(getFs(), path);
   }
 
   @Override
